reorganized constructors &c. made some things into pointers, updated species stuff
