+++ src/AtomicAndPhysicalConstants.jl
@@ -7,8 +7,14 @@ using HTTP
 using JSON
 using Unitful
 
+# kind enum stores the kind of particle
+# NULL is for null species (placeholder species)
+@enumx Kind ATOM HADRON LEPTON PHOTON NULL
+
+include("helpers.jl")
 include("defaults.jl")
 include("constants_struct.jl")
+include("species.jl")
 include("types.jl")
 include("apc_units.jl")
 include("2002_constants.jl")
@@ -18,7 +24,7 @@ include("2014_constants.jl")
 include("2018_constants.jl")
 include("2022_constants.jl")
 include("subatomic_species.jl")
-include("constdef.jl")
+include("setconst.jl")
 # include("constructors.jl")
 # include("isotopes.jl")
 # include("functions.jl")
